Fix handling for single page results
The API always returns data about the entire result set. This data is
used to set the pages and total items count for the PagedResponse. If a
"page" is provided as a query parameter, the PagedResponse constructor
now ignores the regular pagination data and makes the response appear to
be a single page.